Add ability to specify width and height of comparison charts
This commit refers to the pathcompare function of plheatmap.py

Additionally, it adds rudimentary error handling for invalid widths
and heights (i.e. those whose product is less than the number of
heatmaps to be included in the chart)

For example, a comparison chart for 12 hoc files in /home/user/hocs/
could be built by calling pathcompare in one of the following ways:
 - pathcompare('/home/user/hocs/', width = 3, height = 4)
 - pathcompare('/home/user/hocs/', 6, 2)
 - pathcompare('/home/user/hocs/', height = 4, width = 4)

However, the following would print an error and default to building a
square chart appropriate for the number of hoc files being compared:
 - pathcompare('/home/user/hocs/')
 - pathcompare('/home/user/hocs/', width = 3)
 - pathcompare('/home/user/hocs/', width = 1, height = 2)

Note that all of the above examples successfully build and display
comparison charts regardless of whether the values for width and
height are valid. In the above case of 12 hoc files being compared,
the default square chart has a width of 4 and a height of 4.